# packages/api/openapi/agents.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibreChat Agents API",
    "version": "0.1.0",
    "description": "The public agent and skill management endpoints of the LibreChat Agents API. This API is in beta and may change."
  },
  "servers": [
    {
      "url": "agents/v1",
      "description": "Relative to where this document is served (`/api/openapi.json`), so it resolves under any deployment base path."
    }
  ],
  "paths": {
    "/agents": {
      "post": {
        "operationId": "createAgent",
        "summary": "Create an agent",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, or a malformed JSON body",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/JsonParseError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, the caller is banned, or the request fails tenant isolation",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/MessageError"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorMessage"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The bound account is being deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionConflict"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Errors normalized by the route use JSON; the final application error controller sends a text body with the text/html media type.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorMessage"
                    }
                  ]
                }
              },
              "text/html": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "An unknown error occurred.",
                    "Processing error in ErrorController."
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listAgents",
        "summary": "List agents",
        "tags": [
          "Agents"
        ],
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, or a malformed JSON body",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/JsonParseError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, the caller is banned, or the request fails tenant isolation",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/MessageError"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorMessage"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The bound account is being deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionConflict"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Errors normalized by the route use JSON; the final application error controller sends a text body with the text/html media type.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Error"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorMessage"
                    }
                  ]
                }
              },
              "text/html": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "An unknown error occurred.",
                    "Processing error in ErrorController."
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oidcBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "An OIDC access token, sent as a bearer token."
      }
    },
    "schemas": {
      "AgentCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "instructions": {
            "type": [
              "string",
              "null"
            ]
          },
          "instructionsPrompt": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "source": {
                "type": "string",
                "const": "native"
              },
              "groupId": {
                "type": "string"
              },
              "selection": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "const": "production"
                      }
                    },
                    "required": [
                      "type"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "const": "exact"
                      },
                      "promptId": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "promptId"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            },
            "required": [
              "source",
              "groupId",
              "selection"
            ],
            "additionalProperties": false
          },
          "avatar": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "filepath": {
                "type": "string"
              },
              "source": {
                "type": "string"
              }
            },
            "required": [
              "filepath",
              "source"
            ]
          },
          "model_parameters": {
            "type": "object",
            "additionalProperties": {}
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills_enabled": {
            "type": "boolean"
          },
          "skill_authoring_enabled": {
            "type": "boolean"
          },
          "skills_scope": {
            "type": "string",
            "enum": [
              "all",
              "selected",
              "none"
            ]
          },
          "memory_scope": {
            "type": "string",
            "enum": [
              "user",
              "agent"
            ]
          },
          "agent_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "from": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  ]
                },
                "to": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  ]
                },
                "description": {
                  "type": "string"
                },
                "edgeType": {
                  "type": "string",
                  "enum": [
                    "handoff",
                    "direct"
                  ]
                },
                "prompt": {
                  "type": "string"
                },
                "excludeResults": {
                  "type": "boolean"
                },
                "promptKey": {
                  "type": "string"
                }
              },
              "required": [
                "from",
                "to"
              ]
            }
          },
          "end_after_tools": {
            "type": "boolean"
          },
          "hide_sequential_outputs": {
            "type": "boolean"
          },
          "stateful_code_sessions": {
            "type": "boolean"
          },
          "stateful_code_environment": {
            "type": "string",
            "enum": [
              "user",
              "agent-user",
              "conversation"
            ]
          },
          "code_environment_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "code_workspace_id": {
            "anyOf": [
              {
                "type": "string",
                "const": ""
              },
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
              }
            ]
          },
          "repositoryInstructions": {
            "type": "string",
            "enum": [
              "prefer",
              "defer",
              "off"
            ]
          },
          "git_identity": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "email": {
                "type": "string",
                "maxLength": 254
              }
            },
            "required": [
              "name",
              "email"
            ]
          },
          "artifacts": {
            "type": "string"
          },
          "recursion_limit": {
            "type": "number"
          },
          "conversation_starters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tool_resources": {
            "type": "object",
            "properties": {
              "image_edit": {
                "type": "object",
                "properties": {
                  "file_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "execute_code": {
                "type": "object",
                "properties": {
                  "file_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "file_search": {
                "type": "object",
                "properties": {
                  "file_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "vector_store_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "context": {
                "type": "object",
                "properties": {
                  "file_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "ocr": {
                "type": "object",
                "properties": {
                  "file_ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "tool_options": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "defer_loading": {
                  "type": "boolean"
                },
                "allowed_callers": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "direct",
                      "code_execution"
                    ]
                  }
                },
                "run_in_background": {
                  "type": "boolean"
                },
                "describe_intent": {
                  "type": "boolean"
                }
              }
            }
          },
          "subagents": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean"
              },
              "allowSelf": {
                "type": "boolean"
              },
              "shareFiles": {
                "type": "boolean"
              },
              "agent_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "graphs": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "minLength": 1
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": {
                      "type": "string",
                      "minLength": 1
                    },
                    "agent_ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "minItems": 1,
                      "maxItems": 32
                    },
                    "edges": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "from": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "minItems": 1
                              }
                            ]
                          },
                          "to": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "minItems": 1
                              }
                            ]
                          },
                          "description": {
                            "type": "string"
                          },
                          "edgeType": {
                            "type": "string",
                            "const": "direct"
                          },
                          "prompt": {
                            "type": "string"
                          },
                          "excludeResults": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "from",
                          "to",
                          "edgeType"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "entry_agent_id": {
                      "type": "string"
                    },
                    "result_agent_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "name",
                    "description",
                    "agent_ids",
                    "edges",
                    "entry_agent_id",
                    "result_agent_id"
                  ]
                }
              }
            }
          },
          "support_contact": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "anyOf": [
                  {
                    "type": "string",
                    "const": ""
                  },
                  {
                    "type": "string",
                    "format": "email"
                  }
                ]
              }
            }
          },
          "category": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "additionalProperties": false
      },
...
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "AccountDeletionConflict": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "const": "ACCOUNT_DELETION_IN_PROGRESS"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "additionalProperties": false
      },
      "MessageError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": false
      },
      "JsonParseError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "const": "Invalid JSON format"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "additionalProperties": false
      }
    }
  }
}